Refactor numpy stub into module
