+++ stubs/__init__.py
@@ -1,0 +1,5 @@
+"""Lightweight stubs for external dependencies used in tests."""
+
+from . import numpy_stub
+
+__all__ = ["numpy_stub"]
