+++ stubs/numpy_stub.py
